Changed when view-change waits for an adequate view
Based on discussion with Ken, the view-change logic now checks for an
adequate next view as soon as MetaWedged is true, and waits for an
adequate view before attempting ragged edge cleanup. Also, fixed a bug
in finish_view_change where last_suspected was not initialized properly,
and made sure register_predicates() was idempotent.

+++ persistent/PersistNoLog.hpp
@@ -47,7 +47,7 @@ namespace ns_persistent {
     }
     ssize_t nWrite = write(fd,buf,size);
     delete buf;
-    if (nWrite != size) {
+    if (nWrite != (ssize_t) size) {
       throw PERSIST_EXP_WRITE_FILE(errno);
     }
     close(fd);
